11/9

+++ src/SymbolManager/Symbol.h
@@ -9,15 +9,16 @@
 #include <string>
 #include <vector>
 #include "Func.h"
+
 using namespace std;
 
 class Symbol {
 public:
     string word;
-    int type; //0->a,1->a[],2->a[][],-1->func
+    short type; //0->a,1->a[],2->a[][],-1->func
     bool con; //1->const,0->var
     Func *func;
-    Symbol(int type, const string &word, bool con, Func *func);
+    Symbol(short type, const string &word, bool con, Func *func);
 };
